pyproject.toml
[project]
name = "to-clickhouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clickhouse-driver>=0.2.9",
    "pika>=1.3.2",
    "prometheus-client>=0.21.1",
    "python-dotenv>=1.1.0",
]
